% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_prob.R
\name{scale_prob}
\alias{scale_prob}
\title{Scale Probabilities}
\usage{
scale_prob(unscaled, target, wt = NULL)
}
\arguments{
\item{unscaled}{Reported probability
of having the attribute of interest.
A numeric vector with values between 0 and 1.}

\item{target}{Benchmark probability.
A number between 0 and 1.}

\item{wt}{Weights to use when calculating
overall unscaled probability from
individual unscaled probabilities. Optional.}
}
\value{
A numeric vector with scaled
probabilities.
}
\description{
Scale probabilities so that they are
consistent with pre-specified benchmarks,
such as totals obtained from a national survey.
}
\examples{
set.seed(10)
p1 <- runif(n = 10)
p2 <- 0.6
p1_scaled <- scale_prob(unscaled = p1,
                        target = p2)
rbind(p1, p1_scaled)
mean(p1)
mean(p1_scaled)

## use tidyverse functions to scale separately
## within age-sex groups
census <- smoothscale::syn_census
survey <- smoothscale::syn_survey
library(dplyr)

census |>
  left_join(survey, by = c("age", "sex")) |>
  mutate(prob_direct = child_labour / all_children) |>
  group_by(age, sex) |>
  mutate(prob_scaled = scale_prob(unscaled = prob_direct,
                                  target = prob_child_labour))
}
\seealso{
\code{\link[=smooth_prob]{smooth_prob()}}
}
