% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_counts.R
\name{smooth_counts}
\alias{smooth_counts}
\title{Smooth counts that are subject to sampling variation}
\usage{
smooth_counts(count, population, prior_cases = 100)
}
\arguments{
\item{count}{Number of people sampled who
experience the outcome in question.}

\item{population}{Number of people sampled.}

\item{prior_cases}{Variable used to control smoothing
when the number of cases is small.}
}
\value{
A numeric vector with smoothed counts.
}
\description{
Take a set of disaggregated counts drawn from
disaggregated populations,
and smooth towards an overall average,
with small counts being smoothed more than large counts.
}
\details{
The easiest way to smooth sub-populations towards
separate averages is to use
[grouped](https://dplyr.tidyverse.org/reference/group_data.html)
data frames. See the example below.

The `prior_cases` argument controls the amount of smoothing.
Its effect is only noticeable when the number of cases
is very small. Larger values for `prior_cases` produce
more smoothing.
}
\examples{
## use synthetic census data
census <- smoothscale::syn_census

## smooth all groups towards the national level
smoothed <- smooth_counts(count = census$child_labour,
                          population = census$all_children)

## a tidyverse-style way of doing the same thing
library(dplyr)
census \%>\%
  mutate(smoothed = smooth_counts(count = child_labour,
                                  population = all_children))

## use tidyverse functions to smooth
## each age-sex group towards a
## different average
census \%>\%
  group_by(age, sex) \%>\%
  mutate(smoothed = smooth_counts(count = child_labour,
                                  population = all_children))
}
