% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_prob.R
\name{smooth_prob}
\alias{smooth_prob}
\title{Smooth Probabilities}
\usage{
smooth_prob(x, size, prior_cases = 100)
}
\arguments{
\item{x}{Number of successes.}

\item{size}{Number of trials.}

\item{prior_cases}{Variable used to
control smoothing. Default is 100.}
}
\value{
A numeric vector.
}
\description{
Given data on numbers of trials and numbers
of successes in multiple subpopulations,
calculate smoothed probabilities of success
in each subpopulation. For instance,
given  numbers of respondents, and numbers of
employed respondents across many small areas,
calculate smoothed probabilities
of being employed for each area. The smoothing
process pulls probabilities towards the overall mean,
pulling probabilities based on small counts
further than probabilities based on large counts.
}
\section{Stratifying}{


It is often appropriate to stratify a population
and to smooth separately within each strata.
For instance, when estimating probabilities of
being employed, it may be appropriate to divide
the population into strata defined by age and sex.

The easiest way to do stratified smoothing
is to use \href{https://dplyr.tidyverse.org/reference/group_data.html}{grouped}
data frames. See below for an example.
}

\section{Mathematical details}{


The model is

\deqn{x_k \sim \text{Binom}(n_k, \pi_k)}

\deqn{\pi_k \sim \text{Beta}(\lambda \nu, (1 - \lambda) \nu)}

\deqn{\lambda \sim \text{Unif}(0, 1)}

\deqn{\nu \sim \text{Cauchy}^+(m)}

where
\itemize{
\item \eqn{x_k} is the number of successes, which is
specified by argument \code{x}
\item \eqn{n_k} is the number of trials, which is
specified by argument \code{size}
\item \eqn{\pi} is the smoothed probability
of success, the return value from \code{smooth_prob()}, and
\item \eqn{m} is a parameter to control smoothing,
which is specified by argument \code{prior_counts}.
}

The "direct" (unsmoothed) estimate of the
probability of success in subpopulation \eqn{k}
is \eqn{x_k / n_k}.

The value for \eqn{m} only has a noticeable effect
on the results when the number of groups, and the
counts within each group, are small.

For details on the model, see the vignette.
}

\examples{
## use synthetic census data
census <- smoothscale::syn_census

## smooth all groups towards the national level
smoothed <- smooth_prob(x = census$child_labour,
                        size = census$all_children)
smoothed

## compare smoothed and unsmoothed ("direct") estimates
unsmoothed <- census$child_labour / census$all_children
rbind(smoothed, unsmoothed)

## use tidyverse functions to smooth
## each age-sex group towards a
## different average
library(dplyr, warn.conflicts = FALSE)
census |>
  group_by(age, sex) |>
  mutate(smoothed = smooth_prob(x = child_labour,
                                size = all_children))
}
