% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothscale-package.R
\docType{package}
\name{smoothscale-package}
\alias{smoothscale}
\alias{smoothscale-package}
\title{Package 'smoothscale'}
\description{
**smoothscale** contains two functions for
simple small area estimation:
- [smooth_count()] deals with sampling error in counts
  by pulling them towards an overall average
- [scale_count()] deals with measurement error in counts
  by scaling them to be consistent with a (reliable)
  overall total
}
\details{
**smoothscale** also contains a synthetic dataset
called [syn_census].
}
\keyword{internal}
