% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_survey.R
\docType{data}
\name{syn_survey}
\alias{syn_survey}
\title{A synthetic survey dataset}
\format{
A \link[tibble:tibble]{tibble} with 4 rows and the following columns:
\itemize{
\item \code{age}: \verb{"5-9" or }"10-14"`
\item \code{sex}: \code{"Female"} or \code{"Male"}
\item \code{total_child_labour}: Estimated number of children involved in
child labour in the whole country
\item \code{total_all_children}: Estimated number of children, including
those not involved in child labour, in the whole country
}
}
\usage{
syn_survey
}
\description{
A synthetic (ie made-up) dataset, illustrating the
sort of ta bulations than can be produced from a survey
file. The synthetic data measure child labour
at the national level
}
\keyword{datasets}
