% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_prob.R
\name{scale_prob}
\alias{scale_prob}
\title{Scale Probabilities}
\usage{
scale_prob(prob_report, prob_target, wt = NULL)
}
\arguments{
\item{prob_report}{Reported probability
of having the attribute of interest.
A numeric vector with values between 0 and 1.}

\item{prob_target}{Benchmark probability.
A number between 0 and 1.}

\item{wt}{Weights to use when calculating
overall reported probability from
individual reported probabilities. Optional.}
}
\value{
A numeric vector with scaled
probabilities.
}
\description{
Scale probabilities so that they are
consistent with pre-specified benchmarks,
such as totals obtained from a national survey.
}
\examples{
## use synthetic census and survey data
census <- smoothscale::syn_census
survey <- smoothscale::syn_survey

## calculate national prevalence to use as benchmark
prob_national <- sum(survey$total_child_labour) /
                   sum(survey$total_all_children)

## calculate direct estimates of area-level probabilities
prob_area <- census$child_labour / census$all_children

## adjust all groups to match benchmark
prob_scaled <- scale_prob(prob_report = prob_area,
                          prob_target = prob_national)

## use tidyverse functions to adjust each
## age-sex group to a different total
library(dplyr)
target <- survey |>
  mutate(prob_national = total_child_labour / total_all_children)

census |>
  left_join(target, by = c("age", "sex")) |>
  mutate(prob_area = child_labour / all_children) |>
  group_by(age, sex) |>
  mutate(prob_scaled = scale_prob(prob_report = prob_area,
                                  prob_target = prob_national))
}
\seealso{
\code{\link[=smooth_prob]{smooth_prob()}}
}
