% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothscale-package.R
\docType{package}
\name{smoothscale-package}
\alias{smoothscale}
\alias{smoothscale-package}
\title{Package 'smoothscale'}
\description{
\strong{smoothscale} contains two functions for
simple small area estimation:
\itemize{
\item \code{\link[=smooth_prob]{smooth_prob()}} deals with sampling error in probabilities
by pulling them towards an overall average
\item \code{\link[=scale_prob]{scale_prob()}} deals with measurement error in probabilities
by scaling them to be consistent with a (reliable)
overall total!
}
}
\details{
\strong{smoothscale} also contains a synthetic dataset
called \link{syn_census}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bayesiandemography.github.io/smoothscale/}
}

}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

Authors:
\itemize{
  \item Tom Wilson \email{advanceddemographicmodelling@gmail.com}
}

}
\keyword{internal}
