% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_census.R
\docType{data}
\name{syn_census}
\alias{syn_census}
\title{A synthetic census dataset}
\format{
A [tibble][tibble::tibble()] with 80 rows and the following columns:

- `area`: Geographical area, numbered from 1 to 20
- `age`: `"5-9" or `"10-14"`
- `sex`: `"Female"` or `"Male"`
- `child_labour`: Number of children involved in
  child labour
- `all_children`: Total number of children, including
  those not involved in child labour
}
\usage{
syn_census
}
\description{
A synthetic (ie made-up) dataset, illustrating the
sort of tabulations than can be produced from a census file.
The synthetic data measure child labour at the
area level.
}
\keyword{datasets}
